new folder for user boards
