+++ frontend/src/app/pages/board-admin/board-admin.component.ts
@@ -1,19 +1,19 @@
 import { Component, OnInit } from '@angular/core';
-import { UserService } from '../_services/user.service';
+import { UserService } from '../../components/_services/user.service';
 
 @Component({
-  selector: 'app-board-secretary',
-  templateUrl: './board-secretary.component.html',
-  styleUrls: ['./board-secretary.component.css'],
+  selector: 'app-board-admin',
+  templateUrl: './board-admin.component.html',
+  styleUrls: ['./board-admin.component.css'],
   standalone: false
 })
-export class BoardSecretaryComponent implements OnInit {
+export class BoardAdminComponent implements OnInit {
   content?: string;
 
   constructor(private userService: UserService) { }
 
   ngOnInit(): void {
-    this.userService.getSecretaryBoard().subscribe({
+    this.userService.getAdminBoard().subscribe({
       next: data => {
         this.content = data;
       },
