+++ frontend/src/app/pages/board-secretary/board-secretary.component.html
