+++ frontend/src/app/pages/board-student/board-student.component.css
